requirements.txt
python-telegram-bot==20.5
openai==0.28
